+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - LSP: You can now search documents and jump to definitions from entities written in import statements.
 - Added `--backtrace` option to `build`, `run`, `test` commands to enable printing backtrace when a runtime error occurs.
 - Added `backtrace` field to `build` and `build.test` section of `fixproj.toml` file. If it is true, backtrace will be enabled by default.
+- Added `--disable-cpu-feature <feature>` option to `build`, `run`, `test` commands to disable specific CPU features. Add `disable_cpu_features` field to `build` and `build.test` section of `fixproj.toml` file.
 
 ### Changed